+++ src/main/java/com/atwoz/member/domain/member/profile/MemberStyles.java
@@ -1,6 +1,6 @@
-package com.atwoz.member.domain.member.profile.vo;
+package com.atwoz.member.domain.member.profile;
 
-import com.atwoz.member.domain.member.profile.MemberStyle;
+import com.atwoz.member.domain.member.profile.vo.Style;
 import com.atwoz.member.exception.exceptions.member.profile.InvalidStyleException;
 import com.atwoz.member.exception.exceptions.member.profile.StyleDuplicateException;
 import com.atwoz.member.exception.exceptions.member.profile.StyleSizeException;
